requirements.txt
antlr4-python3-runtime==4.7.1
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
bibtexparser==1.1.0
boto==2.49.0
cffi==1.11.5
colorlog==3.1.4
cryptography==2.4.2
cycler==0.10.0
decorator==4.3.2
future==0.17.1
idna==2.7
importlib-metadata==0.18
jsonschema==3.0.1
kiwisolver==1.0.1
matplotlib==3.0.2
meson==0.49.2
more-itertools==7.2.0
networkx==2.2
numpy==1.16.1
packaging==19.0
pluggy==0.12.0
py==1.8.0
pycparser==2.19
pydeps==1.7.1
pydot==1.4.1
pygraphviz==1.5
PyMySQL==0.9.2
pyparsing==2.3.1
pyperclip==1.7.0
pyrsistent==0.14.11
pytest==5.0.1
python-dateutil==2.8.0
six==1.11.0
stdlib-list==0.5.0
wcwidth==0.1.7
z3==0.2.0
z3-solver==4.8.0.0
zipp==0.5.2
